server/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cartesia>=1.4.0",
    "fastapi>=0.115.11",
    "google-genai>=1.7.0",
    "python-dotenv>=1.0.1",
    "uuid>=1.30",
    "uvicorn>=0.34.0",
]
